DataX启动命令修改:   1. python datax.py -f job.json  # 增加-f选项，之后跟job配置文件；   2. python datax.py -e "base64字符串" # 增加-e选项，之后跟job配置内容，该JSON内容需要转化为base64字符创。 hdfsReader: 增加过滤hive临时文件（.hive-staging_hive）； HdfsWriter:   1. writeMode增加overwrite选项，   2. 同时增加tempDirectory临时根目录，数据文件会先写到临时目录，写入成功后，最后在mv到正式目录；   3. hdfs: 修改路径分隔符固定为linux分隔符。 MongoDBReader:   1. column增加value属性，以支持常量（value）;   2. 修复array支持问题，增加ArrayList判断；   3. mongo-java-driver升级到3.10.2；   4. 修复源表无数据时，程序执行异常的问题。 MongoDBWriter:   1. 修复array支持问题，增加ArrayList判断；   2. mongo-java-driver升级到3.10.2。 MySQl: jdbc驱动器升级为8.0.18； MySQLWriter: 增加GEOMETRY支持。

+++ core/src/main/java/com/alibaba/datax/core/statistics/container/communicator/taskgroup/AbstractTGContainerCommunicator.java
@@ -29,7 +29,8 @@ public abstract class AbstractTGContainerCommunicator extends AbstractContainerC
 
     public AbstractTGContainerCommunicator(Configuration configuration) {
         super(configuration);
-        this.jobId = configuration.getInt(
+        // fix bug: getInt => getLong
+        this.jobId = configuration.getLong(
                 CoreConstant.DATAX_CORE_CONTAINER_JOB_ID);
         super.setCollector(new ProcessInnerCollector(this.jobId));
         this.taskGroupId = configuration.getInt(
